starting migration script
