+++ pkg/envparser/createFilesFolder.go
@@ -48,57 +48,3 @@ func CreateDefaultEnvs(envName *string) error {
 	}
 	return nil
 }
-
-// TODO: write function to migrate the pm environemnt and pm collection
-/*
-
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-)
-
-// sample file for env
-
-type EnvValues struct {
-	Key     string `json:"key"`
-	Value   string `json:"value"`
-	Enabled bool   `json:"enabled"`
-}
-
-type Environment struct {
-	Name   string      `json:"name"`
-	Values []EnvValues `json:"values"`
-}
-
-func main() {
-	var env Environment
-	jsonByteVal, err := os.ReadFile("./globals.json")
-	if err != nil {
-		fmt.Println("error occured while opening the json file", err)
-	}
-	err = json.Unmarshal(jsonByteVal, &env)
-	if err != nil {
-		fmt.Println("error occured while unmarshalling the file", err)
-	}
-
-	fmt.Println("Key = ", env.Values[0].Key)
-	fmt.Println("Value \u2713 =", env.Values[0].Value)
-
-	// hulak migrate "./globals.json"
-  // only accept 1 argument and determine
-  // what is it; if  postman envFile Run env otherwise run postman collection
-  // automatically check whether it' postman file
-  // postman has name that says where it's coming from
-
-
-	// if name is empty ""  or name == "globals" then migrate things to global
-	// otherwise a name in pm json file should create a new env file with the exact name if the env file does not exists
-	// if the name in json exists in the env folder there is no need to create it, just migrate
-	// Existing function to create folder and file for the environment
-	// If it's globals then push this into global.env
-	// Otherwise just create then same environment as the name
-}
-*/
